refactor(drive): pull per-mode count executors out of drive_dispatcher.rs
`drive_dispatcher.rs` had two distinct responsibilities glued
together at ~930 LoC: the per-mode `impl Drive` executors
(`execute_document_count_total_no_proof`,
`execute_document_count_per_in_value_no_proof`, etc.) AND the
top-level dispatcher (`execute_document_count_request`) that
routes among them. Physical split:

- `drive_dispatcher.rs` (~470 LoC) — top-level dispatcher,
  `DocumentCountRequest` / `DocumentCountResponse` types,
  `where_clauses_from_value` / `order_clauses_from_value` CBOR
  decoders.
- `executors.rs` (~480 LoC, new) — all six `execute_document_count_*`
  methods on `impl Drive` plus the
  `read_primary_key_count_tree` helper.

Both files now read top-to-bottom in one mental model. The
dispatcher comment block at the top points at the executors
module so the relationship is discoverable. No behavior change;
all 39 rs-drive count + 51 drive-abci v1 + 7 SDK fetch tests
still pass.
